requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymysql==1.1.0
scikit-learn==1.4.0
numpy==1.26.3
PyPDF2==3.0.1
pdfminer.six==20221105
